+++ packages/filter/test/BUILD.bazel
@@ -1,0 +1,20 @@
+package(default_visibility = ["//visibility:public"])
+
+load("//tools/ts_project:index.bzl", "ts_project")
+load("@npm//@bazel/jasmine:index.bzl", "jasmine_node_test")
+
+ts_project(
+    name = "test_lib",
+    testonly = True,
+    srcs = glob(["*.spec.ts"]),
+    deps = [
+        "//packages/database",
+        "//packages/filter",
+        "@npm//@types/jasmine",
+    ],
+)
+
+jasmine_node_test(
+    name = "test",
+    srcs = [":test_lib"],
+)
